Move select, switch, and spinner components out of components/common.js
